--- a/stan/math/prim/mat/functor/map_rect.hpp
+++ b/stan/math/prim/mat/functor/map_rect.hpp
@@ -4,6 +4,7 @@
 #include <stan/math/prim/arr/err/check_matching_sizes.hpp>
 #include <stan/math/prim/mat/fun/dims.hpp>
 #include <stan/math/prim/mat/fun/typedefs.hpp>
+#include <stan/math/prim/scal/meta/return_type.hpp>
 
 #define STAN_REGISTER_MAP_RECT(CALLID, FUNCTOR)