--- a/stan/math/rev/core/autodiffstackstorage.hpp
+++ b/stan/math/rev/core/autodiffstackstorage.hpp
@@ -84,7 +84,6 @@ namespace math {
  * [2] https://github.com/stan-dev/math/pull/826
  * [3]
  * http://discourse.mc-stan.org/t/potentially-dropping-support-for-older-versions-of-apples-version-of-clang/3780/
- * [4] https://github.com/stan-dev/math/pull/1135
  */
 template <typename ChainableT, typename ChainableAllocT>
 struct AutodiffStackSingleton {
@@ -116,9 +115,7 @@ struct AutodiffStackSingleton {
   explicit AutodiffStackSingleton(AutodiffStackSingleton_t const &) = delete;
   AutodiffStackSingleton &operator=(const AutodiffStackSingleton_t &) = delete;
 
-  static inline constexpr AutodiffStackStorage &instance() {
-    return *instance_;
-  }
+  static STAN_THREADS_DEF AutodiffStackStorage *instance_;
 
  private:
   static bool init() {
@@ -129,8 +126,7 @@ struct AutodiffStackSingleton {
     return false;
   }
 
-  static STAN_THREADS_DEF AutodiffStackStorage *instance_;
-  const bool own_instance_;
+  bool own_instance_;
 };
 
 template <typename ChainableT, typename ChainableAllocT>