--- a/stan/math/rev/core/build_vari_array.hpp
+++ b/stan/math/rev/core/build_vari_array.hpp
@@ -20,7 +20,7 @@ namespace math {
 template <int R, int C>
 vari** build_vari_array(const Eigen::Matrix<var, R, C>& x) {
   vari** x_vi_
-      = ChainableStack::instance().memalloc_.alloc_array<vari*>(x.size());
+      = ChainableStack::instance_->memalloc_.alloc_array<vari*>(x.size());
   for (int i = 0; i < x.size(); ++i) {
     x_vi_[i] = x(i).vi_;
   }