--- a/stan/math/rev/core/chainable_alloc.hpp
+++ b/stan/math/rev/core/chainable_alloc.hpp
@@ -16,7 +16,7 @@ namespace math {
 class chainable_alloc {
  public:
   chainable_alloc() {
-    ChainableStack::instance().var_alloc_stack_.push_back(this);
+    ChainableStack::instance_->var_alloc_stack_.push_back(this);
   }
   virtual ~chainable_alloc() {}
 };