--- a/stan/math/rev/core/empty_nested.hpp
+++ b/stan/math/rev/core/empty_nested.hpp
@@ -10,7 +10,7 @@ namespace math {
  * Return true if there is no nested autodiff being executed.
  */
 static inline bool empty_nested() {
-  return ChainableStack::instance().nested_var_stack_sizes_.empty();
+  return ChainableStack::instance_->nested_var_stack_sizes_.empty();
 }
 
 }  // namespace math