--- a/stan/math/rev/core/nested_size.hpp
+++ b/stan/math/rev/core/nested_size.hpp
@@ -8,8 +8,8 @@ namespace stan {
 namespace math {
 
 static inline size_t nested_size() {
-  return ChainableStack::instance().var_stack_.size()
-         - ChainableStack::instance().nested_var_stack_sizes_.back();
+  return ChainableStack::instance_->var_stack_.size()
+         - ChainableStack::instance_->nested_var_stack_sizes_.back();
 }
 
 }  // namespace math