--- a/stan/math/rev/core/recover_memory.hpp
+++ b/stan/math/rev/core/recover_memory.hpp
@@ -20,13 +20,13 @@ static inline void recover_memory() {
     throw std::logic_error(
         "empty_nested() must be true"
         " before calling recover_memory()");
-  ChainableStack::instance().var_stack_.clear();
-  ChainableStack::instance().var_nochain_stack_.clear();
-  for (auto &x : ChainableStack::instance().var_alloc_stack_) {
+  ChainableStack::instance_->var_stack_.clear();
+  ChainableStack::instance_->var_nochain_stack_.clear();
+  for (auto &x : ChainableStack::instance_->var_alloc_stack_) {
     delete x;
   }
-  ChainableStack::instance().var_alloc_stack_.clear();
-  ChainableStack::instance().memalloc_.recover_all();
+  ChainableStack::instance_->var_alloc_stack_.clear();
+  ChainableStack::instance_->memalloc_.recover_all();
 }
 
 }  // namespace math