--- a/stan/math/rev/core/set_zero_all_adjoints.hpp
+++ b/stan/math/rev/core/set_zero_all_adjoints.hpp
@@ -12,9 +12,9 @@ namespace math {
  * Reset all adjoint values in the stack to zero.
  */
 static void set_zero_all_adjoints() {
-  for (auto &x : ChainableStack::instance().var_stack_)
+  for (auto &x : ChainableStack::instance_->var_stack_)
     x->set_zero_adjoint();
-  for (auto &x : ChainableStack::instance().var_nochain_stack_)
+  for (auto &x : ChainableStack::instance_->var_nochain_stack_)
     x->set_zero_adjoint();
 }