--- a/stan/math/rev/core/set_zero_all_adjoints_nested.hpp
+++ b/stan/math/rev/core/set_zero_all_adjoints_nested.hpp
@@ -19,17 +19,17 @@ static void set_zero_all_adjoints_nested() {
     throw std::logic_error(
         "empty_nested() must be false before calling"
         " set_zero_all_adjoints_nested()");
-  size_t start1 = ChainableStack::instance().nested_var_stack_sizes_.back();
+  size_t start1 = ChainableStack::instance_->nested_var_stack_sizes_.back();
   // avoid wrap with unsigned when start1 == 0
   for (size_t i = (start1 == 0U) ? 0U : (start1 - 1);
-       i < ChainableStack::instance().var_stack_.size(); ++i)
-    ChainableStack::instance().var_stack_[i]->set_zero_adjoint();
+       i < ChainableStack::instance_->var_stack_.size(); ++i)
+    ChainableStack::instance_->var_stack_[i]->set_zero_adjoint();
 
   size_t start2
-      = ChainableStack::instance().nested_var_nochain_stack_sizes_.back();
+      = ChainableStack::instance_->nested_var_nochain_stack_sizes_.back();
   for (size_t i = (start2 == 0U) ? 0U : (start2 - 1);
-       i < ChainableStack::instance().var_nochain_stack_.size(); ++i) {
-    ChainableStack::instance().var_nochain_stack_[i]->set_zero_adjoint();
+       i < ChainableStack::instance_->var_nochain_stack_.size(); ++i) {
+    ChainableStack::instance_->var_nochain_stack_[i]->set_zero_adjoint();
   }
 }