--- a/stan/math/rev/core/start_nested.hpp
+++ b/stan/math/rev/core/start_nested.hpp
@@ -11,13 +11,13 @@ namespace math {
  * can find it.
  */
 static inline void start_nested() {
-  ChainableStack::instance().nested_var_stack_sizes_.push_back(
-      ChainableStack::instance().var_stack_.size());
-  ChainableStack::instance().nested_var_nochain_stack_sizes_.push_back(
-      ChainableStack::instance().var_nochain_stack_.size());
-  ChainableStack::instance().nested_var_alloc_stack_starts_.push_back(
-      ChainableStack::instance().var_alloc_stack_.size());
-  ChainableStack::instance().memalloc_.start_nested();
+  ChainableStack::instance_->nested_var_stack_sizes_.push_back(
+      ChainableStack::instance_->var_stack_.size());
+  ChainableStack::instance_->nested_var_nochain_stack_sizes_.push_back(
+      ChainableStack::instance_->var_nochain_stack_.size());
+  ChainableStack::instance_->nested_var_alloc_stack_starts_.push_back(
+      ChainableStack::instance_->var_alloc_stack_.size());
+  ChainableStack::instance_->memalloc_.start_nested();
 }
 
 }  // namespace math