--- a/stan/math/rev/core/vari.hpp
+++ b/stan/math/rev/core/vari.hpp
@@ -56,14 +56,14 @@ class vari {
    * @param x Value of the constructed variable.
    */
   explicit vari(double x) : val_(x), adj_(0.0) {
-    ChainableStack::instance().var_stack_.push_back(this);
+    ChainableStack::instance_->var_stack_.push_back(this);
   }
 
   vari(double x, bool stacked) : val_(x), adj_(0.0) {
     if (stacked)
-      ChainableStack::instance().var_stack_.push_back(this);
+      ChainableStack::instance_->var_stack_.push_back(this);
     else
-      ChainableStack::instance().var_nochain_stack_.push_back(this);
+      ChainableStack::instance_->var_nochain_stack_.push_back(this);
   }
 
   /**
@@ -123,7 +123,7 @@ class vari {
    * @return Pointer to allocated bytes.
    */
   static inline void* operator new(size_t nbytes) {
-    return ChainableStack::instance().memalloc_.alloc(nbytes);
+    return ChainableStack::instance_->memalloc_.alloc(nbytes);
   }
 
   /**