--- a/test/unit/math/mix/scal/meta/broadcast_array_test.cpp
+++ b/test/unit/math/mix/scal/meta/broadcast_array_test.cpp
@@ -1,7 +1,6 @@
+#include <stan/math/mix/scal.hpp>
+
 #include <gtest/gtest.h>
-#include <stan/math/fwd/core/fvar.hpp>
-#include <stan/math/rev/core/var.hpp>
-#include <stan/math/prim/scal/meta/broadcast_array.hpp>
 #include <vector>
 
 TEST(foo, bar) {