… stan/math/rev/mat/fun/cov_exp_quad.hpp
… stan/math/rev/mat/fun/dot_product.hpp
… stan/math/rev/mat/fun/dot_self.hpp
… stan/math/rev/mat/fun/gp_exp_quad_cov.hpp
… stan/math/rev/mat/fun/gp_periodic_cov.hpp
… stan/math/rev/mat/fun/log_determinant.hpp
… stan/math/rev/mat/fun/log_determinant_spd.hpp
… stan/math/rev/mat/fun/mdivide_left.hpp
… stan/math/rev/mat/fun/mdivide_left_ldlt.hpp
… stan/math/rev/mat/fun/mdivide_left_spd.hpp
… stan/math/rev/mat/fun/mdivide_left_tri.hpp
… stan/math/rev/mat/fun/multiply.hpp
… stan/math/rev/mat/fun/multiply_lower_tri_self_transpose.hpp
… stan/math/rev/mat/fun/ordered_constrain.hpp
… stan/math/rev/mat/fun/positive_ordered_constrain.hpp
… stan/math/rev/mat/fun/sd.hpp
… stan/math/rev/mat/fun/simplex_constrain.hpp
… stan/math/rev/mat/fun/softmax.hpp
… stan/math/rev/mat/fun/squared_distance.hpp
… stan/math/rev/mat/fun/sum.hpp
… stan/math/rev/mat/fun/tcrossprod.hpp
… stan/math/rev/mat/fun/unit_vector_constrain.hpp
… stan/math/rev/mat/fun/variance.hpp
… stan/math/rev/mat/functor/adj_jac_apply.hpp
… stan/math/rev/mat/functor/algebra_solver.hpp
… stan/math/rev/mat/functor/map_rect_concurrent.hpp
… stan/math/rev/scal/meta/operands_and_partials.hpp
… test/unit/math/prim/mat/functor/num_threads_test.cpp
… test/unit/math/prim/mat/functor/utils_threads.hpp
… test/unit/math/rev/arr/err/check_bounded_test.cpp