## Summary

This PR reapplies the changes of the PR faster TLS v5 which was reverted from develop after performance tests issues. See #1244

In short, this PR: 
- reverts the revert from #1244, reapplying the fix for threading on Windows
- reverts 2 commits in the TLSv5 branch, effectively goes back to using pointer access to the instance

First, lets see what the perf tests show now. If this wont be fine, we should try ifdefing the autodiffstackstorage.hpp with a more clean version of something like [this](https://gist.github.com/rok-cesnovar/07af3630e15b06d28208bc857cdae199)

## Tests

/

## Side Effects

The side effects to TLSv5 are kind of why we are doing this.

## Performance Results

| OS | compiler | threading | number of threads | schools-4 - develop | schools-4 - tls | schools-4 % improvement | warfarin - develop | warfarin - tls | warfarin % improvement | 
|-----|---------|-----------|-------------------|---------|--------------|---|---|------|---|
| linux | g++ 5.x.x  stdlibc++ |  no | n/a |  10.1534s | 10.1198s | 0.3% | 256.7028s | 238.0342s | 7.8% |
| linux | g++ 5.x.x  stdlibc++ |  yes | 1 |  11.7442s | 10.5511s | 11.3% | 321.1362s | 246.0260s | 30.5% |
| linux | g++ 5.x.x  stdlibc++ |  yes | 2 |  11.4333s | 10.2273s | 11.8% | 291.3092s | 257.3849s | 13.2% |
| linux | g++ 5.x.x  stdlibc++ |  yes | 4 |  11.2953s | 10.0267s | 12.7% | 204.3419s | 175.2958s | 16.6% |
| mac | clang++ 6.x.x  libc++ |  no | n/a |  10.0808s | 10.3644s | -2.7% | 281.9994s | 276.4570s | 2.0% |
| mac | clang++ 6.x.x  libc++ |  yes | 1 |  15.6746s | 10.3569s | 51.3% | 414.8224s | 291.6971s | 42.2% |
| mac | clang++ 6.x.x  libc++ |  yes | 2 |  14.9762s | 10.0897s | 48.4%. | 210.5288s | 156.2712s | 34.7% |
| mac | clang++ 6.x.x  libc++ |  yes | 4 |  14.8488s | 9.9837s | 48.7% | 128.4796s | 94.3915s | 36.1% |
| windows | g++ 4.9.3 |  no | n/a |  9.1471s | 9,1897s | -0.4% | 318.881s | 318,86s | -0.006% |
| windows | g++ 4.9.3 |  yes | 1 |  / | 24.8642s | / | / | 452.852s | / |
| windows | g++ 4.9.3 |  yes | 2 |  / | 23.0454s | / | / | 224.1456s | / |
| windows | g++ 4.9.3 |  yes | 4 | /  | 23,7629s | / | / | 151.7429s | / |