@seantalts 2019-05-18
Not until it’s merged - we have PRs just running on the stat comp
benchmarks suite since it’s shorter and more accepted as meaningful.

We’ll have to run it ourselves, but the scripts should make that relatively
easy other than finding the idle machines.

On Sat, May 18, 2019 at 08:24 Rok Češnovar <notifications@github.com> wrote:

> @seantalts <https://github.com/seantalts> will Jenkins run the entire
> performance tests suite on this PR?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/stan-dev/math/pull/1245?email_source=notifications&email_token=AAGET3HCJZ7MJM2DIPWOTATPV7YPRA5CNFSM4HN2F44KYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODVWNWSA#issuecomment-493673288>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAGET3FBBKU6ZYA24MBWRCDPV7YPRANCNFSM4HN2F44A>
> .
>