@rok-cesnovar 2019-05-18
> We’ll have to run it ourselves

Just to clarify. This means: 
```
git clone --recursive https://github.com/stan-dev/performance-tests-cmdstan
cd performance-tests-cmdstan
./compare-git-hashes.sh example-models/bugs_examples/vol2/schools/ develop d013e55 false false
```
Or something else?