@wds15 2019-05-18
I did run things and got on a Linux RHEL 7.5 box with gcc 6.3.1 these timings:

```
running things a bit shorter:
./example-models/bugs_examples/vol2/schools/schools-4 method=sample num_warmup=250 num_samples=250 data file=example-models/bugs_examples/vol2/schools/schools.data.R random seed=1234

develop-good.log:               267.42 seconds (Total) ## this is before TLSv5 merge
develop-bad.log:               277.86 seconds (Total) ## this is just after TLSv5 merge
develop-v6.log:               262.33 seconds (Total) ## this is for this PR
```

So it looks good to me!