@seantalts 2019-05-18
Rok did that work? The math hash should go in the 3rd slot there:
```
./compare-git-hashes.sh example-models/bugs_examples/vol2/schools/ develop develop false d013e55
```

it's `compare-git-hashes.sh <test_dir> <cmdstan hash/branch of first run> <cmdstan hash of second run> <stan hash of 2nd run> <math hash of 2nd run>` where `false` just means use whatever the cmdstan hash has listed for that submodule.