@rok-cesnovar 2019-05-18
Havent tried yet, just copied pasted from your post on discourse (and mixed up the arguments). Thanks for the info on the arguments. Going to run this on Windows & Ubuntu system.