@seantalts 2019-05-18
Honestly I'm still worried we reverted the wrong PR, or at least that there's another PR out there that is also causing a dip in performance. I'll try to run some git bisect on that soon... Were there any other candidates we were worried about? I know it was suspicious the ODE one seemed to be the first one to cause the failure...