@rok-cesnovar 2019-05-18
You mentioned the ODE, I am not aware of any other. I have only done the git bisect with the schools-4 model as that was the model mentioned. Doing git bisect with more models at a time seemed like a hassle. It seems that the amra model is also returing weird performance results right? At least if I am judging from #1233