@rok-cesnovar 2019-05-18
> it doesn't have anything to do with the TLS stuff right?

Nope. At least as far as I know.