@seantalts 2019-05-18
None of our Macs are on Mojave :sweat_smile: