@SteveBronder 2019-05-19
My work mac is Mojave, if you send me a link to what you want tested I can do it tmrw