@wds15 2019-05-24
I am a bit confused as to how this PR can move forward.

- Who is going to review it? If I were to do that then all is fine (I explain below)
- Why do we still need Mojave? I have Mojave, but why is it needed for what?
- The PR as it is right now addresses the performance regression which was found for the Linux+gcc+schools model constellation. For this setup I have shown above that the performance regression goes away with the code in this PR.

BTW, if I recall right then clang compilers speed up with this changes set a little bit for the single-core case if I am not mistaken... and one more BTW... the performance reports are great to have, but would be better to also have them say what compiler on what OS with what CPU it ran on. I am not saying to expand the tests, just adding that doc helps, I think.