@syclik 2019-05-24
@wds15, I can review it. I agree with all your points below. They were the
same points I made when you originally submitted the PR.

Between the two of us, we can be the "honest broker" and collect the
results accurately and request the right things to be measured. We'll need
to look at: hardware, compiler version, os version, git hash, and relative
computational results.

One other thing we can do is try to simplify the test case so we can
observe these results without needing to go all the way through CmdStan. It
still wouldn't surprise me if there was an interaction with linking that we
didn't know about until now.

Does that sound like it'd work? Wed first define what needs to be measured.

On Fri, May 24, 2019 at 7:27 AM wds15 <notifications@github.com> wrote:

> I am a bit confused as to how this PR can move forward.
>
>    - Who is going to review it? If I were to do that then all is fine (I
>    explain below)
>    - Why do we still need Mojave? I have Mojave, but why is it needed for
>    what?
>    - The PR as it is right now addresses the performance regression which
>    was found for the Linux+gcc+schools model constellation. For this setup I
>    have shown above that the performance regression goes away with the code in
>    this PR.
>
> BTW, if I recall right then clang compilers speed up with this changes set
> a little bit for the single-core case if I am not mistaken... and one more
> BTW... the performance reports are great to have, but would be better to
> also have them say what compiler on what OS with what CPU it ran on. I am
> not saying to expand the tests, just adding that doc helps, I think.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/stan-dev/math/pull/1245?email_source=notifications&email_token=AADH6F2UBBXJLI2N3EMCS4LPW7GI7A5CNFSM4HN2F44KYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODWFAA6Y#issuecomment-495583355>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AADH6F3XVGXXOVXBZAAHU3TPW7GI7ANCNFSM4HN2F44A>
> .
>