@wds15 2019-05-24
@syclik Thanks for reviewing. So what do you think about my suggestion about what to measure? I am pasting the results I posted from above here again:

I did run things and got on a Linux RHEL 7.5 box with gcc 6.3.1 these timings:

running things a bit shorter:
./example-models/bugs_examples/vol2/schools/schools-4 method=sample num_warmup=250 num_samples=250 data file=example-models/bugs_examples/vol2/schools/schools.data.R random seed=1234

```
develop-good.log:               267.42 seconds (Total) ## this is before TLSv5 merge
develop-bad.log:               277.86 seconds (Total) ## this is just after TLSv5 merge
develop-v6.log:               262.33 seconds (Total) ## this is for this PR
```

So it looks good to me!

EDIT: The CPU there is Intel(R) Xeon(R) CPU E5-2670 v2 @ 2.50GHz

So what I did is to run the schools model on this Linux+gcc system (this is exactly the problematic case)

1. before the TLSv5 merge... things are good
2. after the TLSv5 merge... things are bad
3. on the TLSv6 PR ... things are good

Do you think this is sufficient? Do we need more? I mean relative to what we already know, we are good, I think.